Revert "Feat/add models"
