--- .gitattributes
@@ -1,1 +1,0 @@
-*.pbxproj merge=union
