+++ .github/workflows/ios.yml
@@ -2,13 +2,7 @@ name: iOS starter workflow
 
 on:
   workflow_dispatch:
-  pull_request:
-    branches:
-      - 'development'
-      - 'main'
-    paths:
-      - '.github/workflows/ios.yml'
-      - '**/*.swift'
+  push:
 
 jobs:
   build:
