--- Moco/Model/Collection.swift
@@ -1,15 +1,0 @@
-//
-//  Collection.swift
-//  Moco
-//
-//  Created by Aaron Christopher Tanhar on 12/10/23.
-//
-
-import Foundation
-
-struct Collection: Identifiable, Equatable {
-    var title = ""
-    var description = ""
-
-    var id = UUID()
-}
