--- Moco/View/DetectionView.swift
@@ -1,63 +1,0 @@
-//
-//  DetectionView.swift
-//  Moco
-//
-//  Created by Aaron Christopher Tanhar on 13/10/23.
-//
-
-import SwiftData
-import SwiftUI
-
-struct DetectionView: View {
-    @Environment(\.modelContext) private var modelContext
-    @Query private var items: [Item]
-
-    var body: some View {
-//        NavigationSplitView {
-//            List {
-//                ForEach(items) { item in
-//                    NavigationLink {
-//                        Text("Item at \(item.timestamp, format: Date.FormatStyle(date: .numeric, time: .standard))")
-//                    } label: {
-//                        Text(item.timestamp, format: Date.FormatStyle(date: .numeric, time: .standard))
-//                    }
-//                }
-//                .onDelete(perform: deleteItems)
-//            }
-//            .toolbar {
-//                ToolbarItem(placement: .navigationBarTrailing) {
-//                    EditButton()
-//                }
-//                ToolbarItem {
-//                    Button(action: addItem) {
-//                        Label("Add Item", systemImage: "plus")
-//                    }
-//                }
-//            }
-//        } detail: {
-//            Text("Select an item")
-//        }
-        HostedViewController()
-            .ignoresSafeArea()
-    }
-
-    private func addItem() {
-        withAnimation {
-            let newItem = Item()
-            modelContext.insert(newItem)
-        }
-    }
-
-    private func deleteItems(offsets: IndexSet) {
-        withAnimation {
-            for index in offsets {
-                modelContext.delete(items[index])
-            }
-        }
-    }
-}
-
-#Preview {
-    DetectionView()
-        .modelContainer(for: Item.self, inMemory: true)
-}
