+++ Moco/View/HomeView.swift
@@ -9,30 +9,55 @@ import SwiftData
 import SwiftUI
 
 struct HomeView: View {
-    @Environment(\.itemViewModel) private var itemViewModel
+    @Environment(\.modelContext) private var modelContext
+    @Query private var items: [Item]
 
     var body: some View {
-        NavigationStack {
-            List {
-                ForEach(itemViewModel.items) { item in
-                    Text(item.name)
-                }.onDelete { indexSet in
-                    for index in indexSet {
-                        itemViewModel.deleteItem(index)
-                    }
-                }
-            }
-            .toolbar {
-                Button("Add") {
-                    itemViewModel.createItem()
-                }
+//        NavigationSplitView {
+//            List {
+//                ForEach(items) { item in
+//                    NavigationLink {
+//                        Text("Item at \(item.timestamp, format: Date.FormatStyle(date: .numeric, time: .standard))")
+//                    } label: {
+//                        Text(item.timestamp, format: Date.FormatStyle(date: .numeric, time: .standard))
+//                    }
+//                }
+//                .onDelete(perform: deleteItems)
+//            }
+//            .toolbar {
+//                ToolbarItem(placement: .navigationBarTrailing) {
+//                    EditButton()
+//                }
+//                ToolbarItem {
+//                    Button(action: addItem) {
+//                        Label("Add Item", systemImage: "plus")
+//                    }
+//                }
+//            }
+//        } detail: {
+//            Text("Select an item")
+//        }
+        HostedViewController()
+            .ignoresSafeArea()
+    }
+
+    private func addItem() {
+        withAnimation {
+            let newItem = Item(timestamp: Date())
+            modelContext.insert(newItem)
+        }
+    }
+
+    private func deleteItems(offsets: IndexSet) {
+        withAnimation {
+            for index in offsets {
+                modelContext.delete(items[index])
             }
         }
     }
 }
 
 #Preview {
-    @State var itemViewModel = ItemViewModel()
-
-    return HomeView().environment(\.itemViewModel, itemViewModel)
+    HomeView()
+        .modelContainer(for: Item.self, inMemory: true)
 }
